requirements.txt
astroid==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.1
pygame==1.9.6
pylint==2.3.1
PyYAML==5.1.1
six==1.12.0
typed-ast==1.4.0
typing==3.7.4
typing-extensions==3.7.4
wrapt==1.11.2
